base-shop/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.4
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
idpay==0.0.1
kombu==5.2.4
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
redis==4.3.3
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
